feat(DatasetMetadata): add DatasetMetadata to the Dataset page
