+++ src/sections/dataset/Dataset.module.scss
@@ -5,3 +5,12 @@
 .container {
   margin: 0.5rem 0;
 }
+
+.tab-container {
+  padding: 1em 0;
+}
+
+.separation-line {
+  margin: 1em 0;
+  border-bottom: 1px solid #dee2e6;
+}
