+++ src/stories/dataset/dataset-metadata/DatasetMetadata.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react'
 import { WithI18next } from '../../WithI18next'
 import { DatasetMetadata } from '../../../sections/dataset/dataset-metadata/DatasetMetadata'
 import { DatasetMetadataBlock } from '../../../dataset/domain/models/Dataset'
-import { MetadataBlockName } from '../../../dataset/domain/models/MetadataBlockName'
+import { DatasetMockData } from '../DatasetMockData'
 
 const meta: Meta<typeof DatasetMetadata> = {
   title: 'Sections/Dataset Page/DatasetMetadata',
@@ -13,48 +13,7 @@ const meta: Meta<typeof DatasetMetadata> = {
 export default meta
 type Story = StoryObj<typeof DatasetMetadata>
 
-const metadataBlocks: DatasetMetadataBlock[] = [
-  {
-    name: MetadataBlockName.CITATION,
-    fields: {
-      persistentId: 'doi:10.5072/FK2/ABC123',
-      alternativePersistentId: 'doi:10.5072/FK2/ABC123',
-      publicationDate: '2021-01-01',
-      citationDate: '2021-01-01',
-      title: 'Dataset Title',
-      author: [
-        {
-          authorName: 'Admin, Dataverse',
-          authorAffiliation: 'Dataverse.org',
-          authorIdentifierScheme: 'ORCID',
-          authorIdentifier: '123456789'
-        },
-        {
-          authorName: 'Owner, Dataverse',
-          authorAffiliation: 'Dataverse.org',
-          authorIdentifierScheme: 'ORCID',
-          authorIdentifier: '123456789'
-        }
-      ]
-    }
-  },
-  {
-    name: MetadataBlockName.GEOSPATIAL,
-    fields: {
-      geographicUnit: 'km',
-      geographicCoverage: [
-        {
-          geographicCoverageCountry: 'United States',
-          geographicCoverageCity: 'Cambridge'
-        },
-        {
-          geographicCoverageCountry: 'United States',
-          geographicCoverageCity: 'Cambridge'
-        }
-      ]
-    }
-  }
-]
+const metadataBlocks: DatasetMetadataBlock[] = DatasetMockData().metadataBlocks
 
 export const Default: Story = {
   render: () => <DatasetMetadata metadataBlocks={metadataBlocks} />
